Even Liber Abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -2,24 +2,24 @@
 
 int main(void)
 {
-    long first = 1, second = 2;
-    long next;
-    long sum = 2; /* Start with 2 because second is even and already added */
+	int first = 1, second = 2;
+	int next;
+	int sum = 2; /* Start with 2 because second is even and already added */
 
-    /* Generate Fibonacci sequence and sum even-valued terms */
-    while (second <= 4000000)
-    {
-        next = first + second;
-        if (next % 2 == 0)
-        {
-            sum += next;
-        }
-        first = second;
-        second = next;
-    }
+	/* Generate Fibonacci sequence and sum even-valued terms */
+	while (second <= 4000000)
+	{
+		next = first + second;
+		if (next % 2 == 0)
+		{
+			sum += next;
+		}
+		first = second;
+		second = next;
+	}
 
-    printf("%ld\n", sum);
+	printf("%d\n", sum);
 
-    return 0;
+	return 0;
 }
